backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy==2.0.25
google-auth==2.27.0
google-auth-oauthlib==1.2.0
google-api-python-client==2.116.0
python-dotenv==1.0.0
pydantic==2.5.3
pydantic-settings==2.1.0
aiosqlite==0.19.0


